chore: Move rollover api to admin
